Dashboard and Google Calendar Pages
Here are the hi-fi wireframes for the Dashboard page and Google Calendar page (labeled Media on website header). Developers should know that users will be redirected to Dashboard page after logging in. Also on media page, google calendar will be embedded in the empty box and links should be provided for the three social media icons.